requirements.txt
opencv-python
ultralytics
python-dotenv
numpy
Pillow
twilio
easyocr
deep-sort-realtime